requirements.txt
google-adk>=1.0.0
google-genai>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
mcp>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
